# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "这是一个基于Go语言的数据处理与分析平台API服务.",
        "title": "数据处理平台 API",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/datasources": {
            "get": {
                "description": "Retrieve a paginated list of data sources",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Get all data sources",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Number of items per page",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data, total, page, pageSize",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new data source configuration to the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Create a new data source",
                "parameters": [
                    {
                        "description": "Data Source Configuration",
                        "name": "datasource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_foldn_bi-go_internal_models.CreateDataSourceInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_foldn_bi-go_internal_models.DataSource"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/datasources/{id}": {
            "get": {
                "description": "Retrieve a specific data source configuration by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Get a data source by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Data Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_foldn_bi-go_internal_models.DataSource"
                        }
                    },
                    "404": {
                        "description": "Data source not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing data source configuration by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Update an existing data source",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Data Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Source Configuration Update",
                        "name": "datasource",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_foldn_bi-go_internal_models.UpdateDataSourceInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_foldn_bi-go_internal_models.DataSource"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Data source not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a data source configuration by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Delete a data source",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Data Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successfully deleted"
                    },
                    "404": {
                        "description": "Data source not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/datasources/{id}/schema": {
            "get": {
                "description": "Retrieve the top-level schema (e.g., list of tables) of a data source",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Get schema of a data source",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Data Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Schema Information",
                        "schema": {}
                    },
                    "400": {
                        "description": "Invalid ID format",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Data source not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error fetching schema",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/datasources/{id}/schema/{entity_name}": {
            "get": {
                "description": "Retrieve the schema (e.g., columns and types) of a specific table or entity",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "datasources"
                ],
                "summary": "Get schema of a specific entity within a data source",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Data Source ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Entity Name (e.g., table name)",
                        "name": "entity_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Entity Schema Information",
                        "schema": {}
                    },
                    "400": {
                        "description": "Invalid ID or entity name",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Data source or entity not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error fetching entity schema",
                        "schema": {
                            "$ref": "#/definitions/internal_api_v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_foldn_bi-go_internal_models.CreateDataSourceInput": {
            "type": "object",
            "required": [
                "name",
                "type"
            ],
            "properties": {
                "dbName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "otherParams": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "postgresql",
                        "mysql",
                        "csv"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_foldn_bi-go_internal_models.DataSourceType"
                        }
                    ]
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_foldn_bi-go_internal_models.DataSource": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "dbname": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "description": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "otherParams": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_foldn_bi-go_internal_models.DataSourceType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_foldn_bi-go_internal_models.DataSourceType": {
            "type": "string",
            "enum": [
                "postgresql",
                "mysql",
                "csv",
                "clickhouse",
                "sqlite"
            ],
            "x-enum-varnames": [
                "PostgreSQL",
                "MySQL",
                "CSV",
                "ClickHouse",
                "SQLite"
            ]
        },
...
        "github_com_foldn_bi-go_internal_models.UpdateDataSourceInput": {
            "type": "object",
            "properties": {
                "dbName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "filePath": {
                    "type": "string"
                },
                "host": {
                    "type": "string"
                },
                "name": {
                    "description": "Use pointers for optional updates",
                    "type": "string"
                },
                "otherParams": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "port": {
                    "type": "string"
                },
                "type": {
                    "enum": [
                        "postgresql",
                        "mysql",
                        "csv",
                        "clickhouse",
                        "sqlite"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_foldn_bi-go_internal_models.DataSourceType"
                        }
                    ]
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "internal_api_v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}